Fixes for flood protection, Require clang on Linux. Some double to float conversions.

+++ p_view.c
@@ -1070,7 +1070,7 @@ void ClientEndServerFrame(edict_t* ent)
 		bobtime *= 4;
 
 	bobcycle = (int)bobtime;
-	bobfracsin = fabsf(sin(bobtime * M_PI));
+	bobfracsin = fabsf((float)sin(bobtime * M_PI));
 
 	// detect hitting the floor
 	P_FallingDamage(ent);
